rtc_addr+4的情况没有考虑到

+++ abstract-machine/am/src/riscv/npc/timer.c
@@ -7,7 +7,7 @@ void __am_timer_init() {
 void __am_timer_uptime(AM_TIMER_UPTIME_T *uptime) {
   uint32_t high = inl(RTC_ADDR+4);
   uint32_t low = inl(RTC_ADDR);
-  uptime->us = (uint64_t)low + (((uint64_t)high) << 32);
+  uptime->us = low + (((uint64_t)high) << 32);
 }
